Empresas already Created

+++ database/migrations/2021_10_24_110244_create_empresas_table.php
@@ -19,7 +19,7 @@ class CreateEmpresasTable extends Migration
             $table->string('Nombre_Corto', 10);
             $table->string('Tipo_de_Sociedad');
             $table->string('Direccion', 200);
-            $table->mediumInteger('Telefono');
+            $table->Integer('Telefono');
             $table->string('Correo_electronico');
             $table->string('Documento_1');
             $table->string('Documento_2');
